Refactored - moved rendering into separate class OthelloWindow
