+++ src/Othello.cpp
@@ -1,19 +1,8 @@
 #include "Othello.hpp"
 #include <iostream>
-#include "gui/ImGuiWrapper.hpp"
 #include "util/define_logger.hpp"
 
-DEFINE_LOGGER( Othello )
-
-#pragma clang diagnostic push
-#pragma ide diagnostic ignored "cert-err58-cpp"
 namespace {
-    const ImColor boardColor{ 20, 62, 28, 255 };
-    const ImColor blackColor{ 0, 0, 0 };
-    const ImColor blackGhostColor{ 0, 0, 0, 100 };
-    const ImColor whiteColor{ 255, 255, 255 };
-    const ImColor whiteGhostColor{ 255, 255, 255, 100 };
-
     std::ostream& operator<<( std::ostream& ostream, const Othello::Captures& captures ) {
         ostream << '[';
         for ( const auto&[x, y] : captures ) {
@@ -22,135 +11,32 @@ namespace {
         return ostream << ']';
     }
 }
-#pragma clang diagnostic pop
 
-Othello::Othello( bool& stayOpen ) : config{
-        .title = "Game",
-        .open = &stayOpen,
-        .flags = ImGuiWindowFlags_NoTitleBar,
-}, boardState{} {
-    boardState[ 3 ][ 3 ] = State::WHITE;
-    boardState[ 3 ][ 4 ] = State::BLACK;
-    boardState[ 4 ][ 3 ] = State::BLACK;
-    boardState[ 4 ][ 4 ] = State::WHITE;
-}
+DEFINE_LOGGER( Othello )
 
-void Othello::render( gui::ImGuiWrapper& imGuiWrapper ) {
-    ImGui::SetNextWindowPosCenter();
-    ImGui::SetNextWindowSize( { 720, 720 }, ImGuiCond_Once );
-    ImGui::PushStyleColor( ImGuiCol_WindowBg, ImU32( boardColor ));
-    struct Finally {
-        ~Finally() {
-            ImGui::PopStyleColor();
-        }
-    } aFinally;
-    imGuiWrapper.window( config, [ this ] {
-        renderGrid();
-        renderPieces();
-        if ( ImGui::IsMouseHoveringWindow()) {
-            auto mouse = ImGui::GetMousePos();
-            auto pos   = ImGui::GetWindowPos();
-            auto size  = ImGui::GetWindowSize();
-
-            // translate mouse into current window coordinates
-            mouse.x -= pos.x;
-            mouse.y -= pos.y;
-
-            auto x = mouse.x / size.x * 8;
-            auto y = mouse.y / size.y * 8;
-
-            if ( x >= 8 || y >= 8 ) return;
-
-            if ( boardState[ x ][ y ] != State::EMPTY ) return;
-
-            const Captures captures = captured( x, y, blackTurn );
-            if ( captures.empty()) return;
-
-            const State newState = blackTurn ? State::BLACK : State::WHITE;
-            drawGhosts( x, y, blackTurn, captures );
-
-            if ( ImGui::IsMouseClicked( 0 )) {
-                placePiece( x, y, blackTurn, captures );
-                blackTurn = !blackTurn;
-            }
-        }
-    } );
-}
-
-void Othello::renderGrid() {
-    auto      windowSize = ImGui::GetWindowSize();
-    auto      drawList   = ImGui::GetWindowDrawList();
-    auto      pos        = ImGui::GetWindowPos();
-    for ( int i          = 1; i < 8; ++i ) {
-        auto x = windowSize.x / 8 * (float) i;
-        auto y = windowSize.y / 8 * (float) i;
-        drawList->AddLine( { pos.x + x, pos.y },
-                           { pos.x + x, pos.y + windowSize.y },
-                           blackColor,
-                           2 );
-        drawList->AddLine( { pos.x, pos.y + y },
-                           { pos.x + windowSize.x, pos.y + y },
-                           blackColor,
-                           2 );
-
-    }
-}
-
-void Othello::renderPieces() {
-    auto      windowSize = ImGui::GetWindowSize();
-    auto      drawList   = ImGui::GetWindowDrawList();
-    auto      pos        = ImGui::GetWindowPos();
-    float     xOffset    = windowSize.x / 16;
-    float     yOffset    = windowSize.y / 16;
-    float     xSize      = windowSize.x / 8;
-    float     ySize      = windowSize.y / 8;
-    for ( int i          = 0; i < boardState.size(); ++i ) {
-        for ( int j = 0; j < boardState.size(); ++j ) {
-            ImColor color{};
-            switch ( boardState[ i ][ j ] ) {
-                case State::EMPTY: continue;
-                case State::WHITE: color = whiteColor;
-                    break;
-                case State::BLACK: color = blackColor;
-                    break;
-            }
-            drawList->AddCircleFilled( { pos.x + xSize * (float) i + xOffset, pos.y + ySize * (float) j + yOffset },
-                                       std::min( xSize / 3, ySize / 3 ), color, 24 );
-        }
-    }
+Othello::Othello() : boardState_{} {
+    boardState_[ 3 ][ 3 ] = State::WHITE;
+    boardState_[ 3 ][ 4 ] = State::BLACK;
+    boardState_[ 4 ][ 3 ] = State::BLACK;
+    boardState_[ 4 ][ 4 ] = State::WHITE;
 }
 
 void Othello::placePiece( int x, int y, bool isBlack, const Captures& captures ) {
-    auto& place = boardState.at( x ).at( y );
+    auto& place = boardState_.at( x ).at( y );
     LOG4CPLUS_DEBUG( GetLogger(),
                      "Placing " << ( isBlack ? "Black" : "White" ) <<
                                 " piece at (" << x << ',' << y << ')' << " Capturing:" << captures );
     const auto& newState = isBlack ? State::BLACK : State::WHITE;
     place = newState;
     for ( const auto[x_, y_]: captures ) {
-        boardState[ x_ ][ y_ ] = newState;
+        boardState_[ x_ ][ y_ ] = newState;
     }
-}
-
-void Othello::drawGhosts( int x, int y, bool black, const Captures& captures ) {
-    auto  windowSize = ImGui::GetWindowSize();
-    auto  drawList   = ImGui::GetWindowDrawList();
-    auto  pos        = ImGui::GetWindowPos();
-    float xOffset    = windowSize.x / 16;
-    float yOffset    = windowSize.y / 16;
-    float xSize      = windowSize.x / 8;
-    float ySize      = windowSize.y / 8;
-    drawList->AddCircleFilled( { pos.x + xSize * (float) x + xOffset, pos.y + ySize * (float) y + yOffset },
-                               std::min( xSize / 3, ySize / 3 ), black ? blackGhostColor : whiteGhostColor, 24 );
-    for ( const auto[x, y]: captured( x, y, black )) {
-        drawList->AddCircleFilled( { pos.x + xSize * (float) x + xOffset, pos.y + ySize * (float) y + yOffset },
-                                   std::min( xSize / 3, ySize / 3 ), black ? blackGhostColor : whiteGhostColor, 24 );
-    }
+    blackTurn = !blackTurn;
 }
 
 std::vector<std::pair<int, int>> Othello::captured( int x, int y, bool isBlack ) const {
     Captures captures;
-    const auto& state = boardState.at( x ).at( y );
+    const auto& state = boardState_.at( x ).at( y );
     if ( state != State::EMPTY ) return captures;
     const State same     = isBlack ? State::BLACK : State::WHITE;
     const State opposite = isBlack ? State::WHITE : State::BLACK;
@@ -159,7 +45,7 @@ std::vector<std::pair<int, int>> Othello::captured( int x, int y, bool isBlack )
     {
         Captures  temp;
         for ( int i = y + 1; i < 8; ++i ) {
-            const State& nextSpot = boardState[ x ][ i ];
+            const State& nextSpot = boardState_[ x ][ i ];
             if ( nextSpot == opposite ) temp.push_back( { x, i } );
             if ( nextSpot == same ) {
                 captures.insert( captures.end(), temp.begin(), temp.end());
@@ -173,7 +59,7 @@ std::vector<std::pair<int, int>> Othello::captured( int x, int y, bool isBlack )
     {
         Captures  temp;
         for ( int i = y - 1; i >= 0; --i ) {
-            const State& nextSpot = boardState[ x ][ i ];
+            const State& nextSpot = boardState_[ x ][ i ];
             if ( nextSpot == opposite ) temp.push_back( { x, i } );
             if ( nextSpot == same ) {
                 captures.insert( captures.end(), temp.begin(), temp.end());
@@ -188,7 +74,7 @@ std::vector<std::pair<int, int>> Othello::captured( int x, int y, bool isBlack )
     {
         Captures  temp;
         for ( int i = x + 1; x < 8; ++i ) {
-            const State& nextSpot = boardState[ i ][ y ];
+            const State& nextSpot = boardState_[ i ][ y ];
             if ( nextSpot == opposite ) temp.push_back( { i, y } );
             if ( nextSpot == same ) {
                 captures.insert( captures.end(), temp.begin(), temp.end());
@@ -202,7 +88,7 @@ std::vector<std::pair<int, int>> Othello::captured( int x, int y, bool isBlack )
     {
         Captures  temp;
         for ( int i = x - 1; i >= 0; --i ) {
-            const State& nextSpot = boardState[ i ][ y ];
+            const State& nextSpot = boardState_[ i ][ y ];
             if ( nextSpot == opposite ) temp.push_back( { i, y } );
             if ( nextSpot == same ) {
                 captures.insert( captures.end(), temp.begin(), temp.end());
@@ -215,7 +101,7 @@ std::vector<std::pair<int, int>> Othello::captured( int x, int y, bool isBlack )
     {
         Captures  temp;
         for ( int i = x + 1, j = y + 1; i < 8 && j < 8; ++i, ++j ) {
-            const State& nextSpot = boardState[ i ][ j ];
+            const State& nextSpot = boardState_[ i ][ j ];
             if ( nextSpot == opposite ) temp.push_back( { i, j } );
             if ( nextSpot == same ) {
                 captures.insert( captures.end(), temp.begin(), temp.end());
@@ -228,7 +114,7 @@ std::vector<std::pair<int, int>> Othello::captured( int x, int y, bool isBlack )
     {
         Captures  temp;
         for ( int i = x - 1, j = y - 1; i >= 0 && j >= 0; --i, --j ) {
-            const State& nextSpot = boardState[ i ][ j ];
+            const State& nextSpot = boardState_[ i ][ j ];
             if ( nextSpot == opposite ) temp.push_back( { i, j } );
             if ( nextSpot == same ) {
                 captures.insert( captures.end(), temp.begin(), temp.end());
@@ -241,7 +127,7 @@ std::vector<std::pair<int, int>> Othello::captured( int x, int y, bool isBlack )
     {
         Captures  temp;
         for ( int i = x - 1, j = y + 1; i >= 0 && j < 8; --i, ++j ) {
-            const State& nextSpot = boardState[ i ][ j ];
+            const State& nextSpot = boardState_[ i ][ j ];
             if ( nextSpot == opposite ) temp.push_back( { i, j } );
             if ( nextSpot == same ) {
                 captures.insert( captures.end(), temp.begin(), temp.end());
@@ -254,7 +140,7 @@ std::vector<std::pair<int, int>> Othello::captured( int x, int y, bool isBlack )
     {
         Captures  temp;
         for ( int i = x + 1, j = y - 1; i < 8 && j >= 0; ++i, --j ) {
-            const State& nextSpot = boardState[ i ][ j ];
+            const State& nextSpot = boardState_[ i ][ j ];
             if ( nextSpot == opposite ) temp.push_back( { i, j } );
             if ( nextSpot == same ) {
                 captures.insert( captures.end(), temp.begin(), temp.end());
@@ -269,7 +155,7 @@ std::vector<std::pair<int, int>> Othello::captured( int x, int y, bool isBlack )
 
 std::pair<int, int> Othello::score() const {
     std::pair<int, int> value{};
-    for ( const auto& cells: boardState )
+    for ( const auto& cells: boardState_ )
         for ( const auto& cell: cells )
             switch ( cell ) {
                 case State::EMPTY: break;
